CreatePlaylist/DeletePlaylist work fine.

+++ src/use_case/delete_playlist/DeletePlaylistInteractor.java
@@ -2,6 +2,8 @@ package use_case.delete_playlist;
 
 import entity.PlaylistFactory;
 
+import javax.swing.*;
+
 public class DeletePlaylistInteractor implements DeletePlaylistInputBoundary {
     final DeletePlaylistDataAccessInterface deletePlaylistDataAccessObject;
     final DeletePlaylistOutputBoundary deletePlaylistPresenter;
@@ -13,10 +15,22 @@ public class DeletePlaylistInteractor implements DeletePlaylistInputBoundary {
     }
     @Override
     public void execute(DeletePlaylistInputData deletePlaylistInputData) {
-        String playlistName = deletePlaylistInputData.getName();
-        deletePlaylistDataAccessObject.delete(playlistName);
+        if (deletePlaylistInputData.getName() == null) {
+            deletePlaylistPresenter.prepareFailView("The playlist name can't be empty");
+        }
+        else if (!deletePlaylistDataAccessObject.existsPlaylistName(deletePlaylistInputData.getName())) {
+            deletePlaylistPresenter.prepareFailView("The playlist does not exist");
+        }
+        else {
+            String playlistName = deletePlaylistInputData.getName();
+            int dialogResult = JOptionPane.showConfirmDialog(null,
+                    "Are you sure you want to say good bye to " + playlistName + "?");
+            if (dialogResult == JOptionPane.YES_OPTION) {
+                deletePlaylistDataAccessObject.delete(playlistName);
 
-        DeletePlaylistOutputData deletePlaylistOutputData = new DeletePlaylistOutputData(playlistName);
-        deletePlaylistPresenter.prepareSuccessView(deletePlaylistOutputData);
+                DeletePlaylistOutputData deletePlaylistOutputData = new DeletePlaylistOutputData(playlistName);
+                deletePlaylistPresenter.prepareSuccessView(deletePlaylistOutputData);
+            } else {deletePlaylistPresenter.prepareFailView("a");}
+        }
     }
 }
